Switch the PHP client to Asciidoctor
Switches the PHP client's doc to being built from Asciidoctor. The
resulting html has quite a few changes spacing changes that aren't
impactful and a couple of small changes that look fine visually. There
are a few changes to auto-generated IDs that we've grown to expect from
asciidoctor. They seem fine to me. Here is the diff:
https://gist.github.com/nik9000/72113b441f682b116b7b41a2b78baed6