Asciidoctor: Fix Elastic's include-tagged
Elastic has has tagged include support in AsciiDoc for years now. Once
we've migrated to Asciidoctor we can start using its native tagged
include support, but while we're in the process of migrating we need to
perfectly emulate Elastic's tagged include in Asciidoctor. We've had
this emulation for a while and this change fixes a bug which caused
empty lines to be throw out. Let's look at an example:

Say you have some file like this:
```
...
    // tag::foo
    String foo = "bar";

    System.err.println("ASDFDAF " + foo);
    // end::foo
...
```

Elastic's AsciiDoc include tagged support would extract the `foo` tag
as:
```
    String foo = "bar";

    System.err.println("ASDFDAF " + foo);
```

The bug in our Asciidoctor include tagged support is that it'd extract
the `foo` tag as:
```
    String foo = "bar";
    System.err.println("ASDFDAF " + foo);
```

Those empty lines can be quite useful in examples! We should preserve
them.