Add redirects to built docs (#674)
This causes us to build a list of redirects in the built-docs repo which
we can then pick up on the public web server. For now, this list is just
copied from the public web server and munged slightly to make it
testable locally.

As a side advantage, when you run `./build_docs --all --open` it'll
include the same redirects that we have on the public web server!