Switch the Infra Monitoring Guide to asciidoctor
Switches the Infrastructure Monitoring Guide to Asciidoctor. There are a
few spacing changes that are noops in HTML and there are a few changes
that look like this:
```
-            <img alt="monitoring-architecture.png" src="monitoring-architecture.png"/>
+            <img alt="monitoring architecture" src="monitoring-architecture.png"/>
```

They switch the alt tag on an image from the name of the image to a
"more English" version of the name of the image. I'm not sure if this is
really an improvement because the alt tag is still being automatically
generated so it isn't going to be particularly descriptive. But I don't
*think* this is a problem either.