self-test: remove last temporary workaround (#706)
Remove the last "temporary workaround" in `html_diff` for issues
building the readme. I fix the readme by using `ifdef` to write
different source text for AsciiDoc vs Asciidoctor. This would be a bad
solution if these things were common in normal books, but the constructs
that need this are when you have snippets of asciidoc inside the
asciidoc. This is a thing that will probably only happen in the readme.