Switch APM Agent for Go to Asciidoctor
Switches the core of the process that builds the docs for the APM Agent
for Go from the no-longer-supports AsciiDoc to the supported
Asciidoctor. The html output is slightly different - mostly spacing
which is generally ignored by the browser. It *does* change the spacing
in front of the go source from a leading eight spaces to a leading tab.
These look the same to me in the browser but copy-and-pasting them comes
out slightly differently.