Build tables of contents with Asciidoctor
We chain the books that make the all of the docs together into a web
site by making tables of content that link the whole thing together.
We make these by building asciidoc files and then building them like
single-page books. This switches from building them with AsciiDoc to
building them with Asciidoctor.