Asciidoctor: BWC for // CONSOLE then callouts
Our backwards compatibility support for "OPEN IN CONSOLE"-style links did
not work with snippets followed by a callout list. That looks like this:

```
[source,js]
----
GET / <1>
----
// CONSOLE
<1> words go here
```

These is actually *super* common and I'm surprised I didn't catch it
before! This adds tests for it so it won't sneak through again.

The reason this didn't work is that our backwards compatibility worked by
transforming the `// CONSOLE` comment in `pass:[// CONSOLE]` and then
catching that when walking the tree. We do this in two passes because our
preprocessor works line by line and reading ahead can cause issues. So we
do the "action at a distance" in a tree processor where it is quite
simple.

The trouble is that the `pass:[// CONSOLE]` construct is *inline* so when
it is immediately followed by the `<1> words go here` line it gets sucked
into the same block which Asciidoctor thinks of as a paragraph. This
causes the parsing for the callout list not to happen, causing all kinds
of trouble.

The fix for this is to introduce a block macro that functions like
`pass:[// CONSOLE]` and to use that instead. This causes the callout list
to be parsed properly.