Switch Python APM Agent docs to Asciidoctor (#723)
Switches the core of the docs generation pipeline from the
no-longer-maintained AsciiDoc project to the actively-maintained
Asciidoctor project. The HTML that it produces isn't 100% the same but
it looks to only differ in three ways:
1. Spacing that the browser will ignore anyway.
2. AsciiDoc was producing `<p></p>` inside of empty table cells.
Asciidoctor does not.
3. In the 1.x docs there are some attribute references that do not
resolve and they are rendered differently by Asciidoctor. Either way,
they should probably be fixed in the 1.x branch of the docs.