Skip post-book actions if we can (#673)
Skips some of the actions that we take after building all branches of a
book if we can get away with skipping them and reworks some of the
logging. This should marginally speed up the `--all` build and and make
it *much* easier to spot which books are rebuilt in which build by
suppressing a lot of the logging noise.