Commit template instead of fetching it (#689)
This stops `build_docs` from fetching the html template used to spruce
up the docs built by docbook to look more like the rest of Elastic's
site. Instead we commit the template directly to the repository and add
a simple shell script to fetch it. I've extracted a promise from the
folks that maintain the template that they'll either open a PR to update
the template or file an issue about it.

This drops support for `--staging` which builds the docs against the
template from Elastic's staging site. I do this because we're honestly
not sure how we'd like to use it and it hasnt worked in CI ever. We'll
work something out when we know what we have to do with staging.