Switch APM Overview to Asciidoctor
Switches the core of the docs build for the APM overview from the
unmaintained AsciiDoc to the actively maintained AsciiDoctor project. It
is marginally faster to build as well. The HTML output is the same
modulo some spacing differences and one real change to the tables on the
`install-and-run` page. The first element on the table gets the
"heading" formatting. Like this:
```
-                <a class="ulink" href="https://www.elastic.co/guide/en/apm/agent/js-base/current/intro.html" target="_top">
-                 Introduction
-                </a>
+                <span class="strong strong">
+                 <strong>
+                  <a class="ulink" href="https://www.elastic.co/guide/en/apm/agent/js-base/current/intro.html" target="_top">
+                   Introduction
+                  </a>
+                 </strong>
+                </span>
```

This is worth looking into. It isn't what we want but I can see *why*
Asciidoctor does it.