Make x-pack tags clickable
This replaces the `x-pack` tags on the titles of pages that are about
elastic licensed features with clickable links that take you to the
subscriptions page.

This is implemented by replacing the section rendering code in docbook
with customized code that includes the link.

This also replaces the background image of the `x-pack` tag with one
that looks "more clickable" and has a hover treatment.

Closes #720