Fix sharing docs built with firebase
Sometimes we want to share the docs that we've built with firebase or a
similar too. I broke that a while back when I turns off `abs_url` when
you build with `--doc` inside docker because I *thought* `abs_url` was
only there to support the built in python web server that we use with
`--open` outside of docker. But it also supports sharing docs with
firebase.

Eventually we'd like something more automatic than firebase but firebase
is what we have right now....