Use more standard rspec for the care admonition
This switches the way we write the tests for the "care admonitions" to a
more rspec-standard style using nested `context`s to describe the input.
It also adds support to the new `convert` shared context for
automatically asserting there nothing was logged during the conversion
process. This is what we want the vast majority of the time so you have
to take some action to disable the automatic assertion, namely `let`ing
a variable to `true`.