Be a little more ruby-standard (#762)
This switches the way we fetch dependencies for ruby from `gem` to
`bundler` which is more standard in the ruby world. This is basically a
noop unless you are doing ruby development. But if you *are* doing ruby
development then you can use standard ruby commands like `bundle exec
rspec` locally which makes the project *much* more approachable for ruby
folks. It isn't authoritative though. Jenkins will still use docker, as
will building the docs.

I've not gone so far as to use rake for ruby stuff yet. I'll get there
in a subsequent PR probably.