Begin moving integration tests to rspec (#765)
This begins the process of moving integration tests from the `Makefile`
in `integtest` to spec in `integtest/spec`. It begins by switching some
of the "expected files" tests and is meant as a pattern to build on.

The interesting thing from an rspec perspective here is that we want to
run each conversion once, before any assertions. Rspec's standard `let`
based building blocks aren't quite right for this. Instead I rely on
careful placement of `before(:context)` blocks. Blocks hooked to
`before(:context)` will run before the context in which they are
defined.

Here is a fairly normal context, asserting that beta[] works as expected:
```
  context 'when the book contains beta[]' do
    def make_input
      write_source 'index.asciidoc', <<~ASCIIDOC
        #{HEADER}
        beta[]

        Words
      ASCIIDOC
    end
    include_context 'convert'

    it 'copies the warning image' do
      expect(dest_file('images/icons/warning.png')).to file_exist
    end
    context 'chapter.html' do
      include_context 'page body'
      it 'includes the warning image' do
        expect(body).to include(
          '<img alt="Warning" src="images/icons/warning.png" />'
        )
      end
      it 'includes the beta text' do
        expect(body).to include(
          'The design and code is less mature than official GA features'
        )
      end
    end
  end
```

There is a fair bit going on so I'll annotate. First we define
`make_input` which is called later to build the source files and returns
the "root" source file that is passed as the argument to `--doc`. Next
we include the `convert` context. This defines the `before(:context)`
hooks that builds a temp directory, calls `make_input` to setup the
source documents, and performs the conversion. It also defines an
`after(:context)` hook to clean up the temp directory. The important bit
here is that you can have `context`s inside of this context but that the
temp directory survives all of them.

Next we make a fairly banal assertion that we copy an image.
Interestingly, the error reporting from rspec's built in `exists`
matcher is total garbage so I make my own and use it here.

Next we make a context called `chapter.html` to make assertions about
the `chapter.html` page. The first thing we do is include the `page
body` context which adds an assertion that the page exists and defines
some `lets` which we can use to assert things about the page. We then
assert some things about the beta widget that should be on that page.