integtest: Implement Asciidoctor dueling in rspec (#805)
Implements a test that compares the output of converting the docs using
Asciidoctor to using AsciiDoc and adds that to all "single book" test
cases. This allows us to entirely move the tests that used a few files
from the Makefile to rspec. Because we write those files to a tmpdir
directly from the rspec file we can now remove the asciidoc files in the
integtest folder.

Also moves the test for the `experimental` admonition into rspec to be
with its brother the `beta` admonition.