Use sparse checkout to speed up building all docs (#830)
* Use sparse checkout to speed up building all docs

This begins to use sparse checkout to speed up building `--all` docs. It
does so in the simplest way possible at this point: instead of checking
out the entire target repo before checking out other repos we checkout
all of the target repo *except* the built docs before checking out the
other repos. This allows us to read the "tracker" file from the built
docs repo to pick which hashes to use for the other repos. We then
checkout the built docs in parallel with the other repos.

The bulk of changed code is around breaking out handling of the
"target" repo into a new class `TargetRepo` and creating a `BaseRepo`
that is shared between `TargetRepo` and `Repo`.

This doesn't save *that* much time locally but I expect it'll be
somewhat more substantial in CI because CI tends to spend more time
pulling changes. In addition, I think we can do better later by using
sparse checkout on the built docs. In most cases we're not actually
rebuilding all of the docs so we don't need to check them all out.

* Words

* Empty ok

* fixup