Fix up-to-date check with asciidoctor
It turns out that we've had a broken up-to-date check for books build
with asciidoctor because we were accidentally passing the "this book is
built with asciidoctor" metadata into git when it was expecting just a
hash. This caused *any* change in a repo to force rebuilding every book
that uses that repo that is built with asciidoctor. Ooops.

The vast, vast, vast majority of this commit is building the
infrastructure needed to test it. That is because we don't have the
infrastructure to use rspec to test building all books more than once.