Allow building docs with an Elastic clone (#846)
When you build docs with `--doc` we attempt to figure out the "edit me"
url. If we can't find an "elastic" clone we would log a warning and move
on. The trouble is that this warning upsets the docs build unless you
pass `--lenient`. This change fixes the problem by instead using the
first remote that it finds if it can't find the Elastic remote. And if
it can't find *any* remotes then it uses `unknown` as the remote. This
prevents the warning.

Closes #845