Add basic test for docs.js
Adds a single, super basic test for docs.js using nodejs, yarn, and
jest. The majority of the complexity here is in getting the nodejs
environment set up in the docker container. We do it similarly to how we
set up ruby:
* Install nodejs with that package manager
* Use a nodejs specific tool to fetch dependencies: yarn
* Lock the dependencies with a lockfile to prevent them from wandering
around when we rebuild the docker image

I chose not to use nvm here because yarn wanted to be installed using
apt and it felt right to install nodejs the same way.

I use the same version of jquery that we use on the public site even if
it is super old.