Drop template.md5 file (#855)
Now that we're not automatically fetching the template there isn't a
nead to store its md5 with every book. Instead we leave a note to users
that they'll have to manually force a rebuild after updating the
template. This saves us a few files which is nice but it allows us to
significantly simplify the code around up-to-date checking.