Fix up-to-date check when switching asciidoctor-ness (#867)
When I added support for asciidoctor I mistakenly caused all books built
by asciidoctor to be considered out of date when there was *any* change
made to any repo that builds them, even if the changes were to unrelated
files.

When I fixed that I mistakenly caused remove asciidoctor-ness from
consideration in the up-to-date check entirely. Programing is like that,
I guess. This fixes that by explicitly checking if asciidoctor-ness
changed.