Teach Kibana link checking about PR features
The docs build has three features when building all books:
1. If you specify `--keep_hash` we reuse the hash of the source repos
for every book that we've built before. This is useful for rebuilding
only the books that we converted to asciidoctor. Or in combination with
`--sub_dir`
2. If you specify `--sub_dir` we use a directory instead of a commit
hash from some source repo. This is mostly useful in combination with
`--keep_hash` because when you use those together your can rebuild the
docs as though you committed the directory that you are subsituting. It
makes for perfect pull request tests.
3. We automatically check all of the links, including links from some
typescript files in the kibana repo.

The trouble is that the kibana repo link checking process didn't
understand `--keep_hash` and `--sub_dir`. This commit changes that. The
bulk of it the change is testing the link checking, which we didn't
really have any of up until now.

Closes #892