Remove edit link from links to bridgeheads
We had a bug where links to floating titles (aka bridgeheads) would
include the `edit_me` link in its contents like this:

```
<a class="xref" href="chapter.html#floater" title="Floateredit">
  Floater
  <a href="https://github.com/elastic/docs/edit/master/index.asciidoc" class="edit_me" title="Edit this page on GitHub" rel="nofollow">
    edit
  </a>
</a>
```

Because this is invalid html it'd cause the `edit` button to float to
the top of the page, and, confusingly, take you to edit the source of
the *link target* rather than the current page. Madness.

This removes the extra link.

Closes #903