Remove legacy saas-release from ESS doc builds (#928)
Last year, we switched to the `release-ms-nn` doc builds for `current`. At the time, we decided to keep `saas-release` around for a while to prevent breaking bookmarks and links for people. We've now been operating under the new milestone release builds for some time and I think it's time to remove the old legacy `saas-release` docs after giving folks a heads-up in the next weekly update. 

Just today, I looked at a recent doc issue that referenced some docs from `saas-release`, but at this point they are at least six or seven months out of date. I think it's time to rip off the bandaid.

@elastic/cloud-writers please approve--we'll discuss tomorrow as well.