Preview: More closely mimick nginx on dir request
When you request a directory on the public web site we redirect your
request to `{path}/index.html`. The preview *was* just serving that
index file instead. This sounds good but breaks relative links in the
`index.html` file. Sadly, we have some of those. This change teaches the
preview to do the same redirect as the public site.