Create branch cleaner for preview (#919)
Creates a fairly simple script that removes the branches created by pull
request builds when the pull requests have closed. The plan is to
schedule this once a day or something in jenkins so it can keep the
`build-docs` repo from growing too large.

The docs preview service will notice the deleted branches and clean them
up on the machines where it is running.