Edit me link improvements (#994)
Makes two changes to the generation of "edit" links for asciidoctor:
1. Prefers the longest matching repository. This helps when repository
is `logtstash` and another is `logstash-docs`.
2. Create a setting to enable overriding `edit_me` links using an
attribute. Historically we've overridden this attribute in cases where
it is no longer needed for Asciidoctor and overriden it *incorrectly*.
But there are some cases where we really do want ot override it.
Specifically for logstash. So this allows us to turn it on when we want
it.