Preview: Add a diff endpoint (#1004)
I've found that it can be kind of annoying trying to find out which
files changed when looking up a docs preview. This adds a page at
`/diff` that contains links to all of the modified pages.

I used chunked transfer and stream parsing the result from git because
that felt like the right thing to do here but it did make the parsing a
bit more complex.

I also chose to use the `-z` flag when calling git because it emits a
format that isn't jumbled when files move. It has many ascii-nul
terminated records which means I couldn't really use any off the shelf
nodejs parsing libraries.