Lookup alternative examples in all subdirs (#1037)
Changes the build so that, when looking for alternative examples, it
looks in all subdirectories. Instead of performing file system
operations for all lookups it indexes all of the examples up front. This
is significantly faster when there are many subdirectories.

While working on this I discovered that we can almost certainly cache
the results of converting the documents for the alternatives. That will
*probably* save a good bit of time. But it isn't 100% simple. So I left
a TODO to investigate it.