Update code examples style
We'd like to customize the style of the code examples. To do that we need
to serve the style locally. To do that we need to serve `code-prettify`
locally. This does that by copying it into `resources/web/prettify` and
modifying it to work properly with parcel and bundling it alone with the
rest of our js.

As a bonus this adds support for SQL which we've started to include in
the ES docs.