Drop animated scrolling
We had animated the scroll when you click on an ID or load a page with
an id to make sure we didn't cover up a "floating" pane at the top of
the page. We don't have such a pane anymore so we can fall back to the
browser's default behavior.

Closes #950