Further modernize web resources (#1055)
This moves from one css and js file per book to one css and js file
globally and drops support for "abs mode" templates which were required
to make demos before we had a fairly easy way to make previews. Dropping
"abs mode" was required to support the global css and js. The global css
and js is nice because it'll allow us to deploy css and js changes
without rebuilding "all" books. This means we'll be able to deploy those
changes much more quickly.

This change also pushes the css file through parcel which minifies it.
We'll follow this change up with one that better organizes the CSS and
parcel will transparently merge them.

This change also enables parcel's server when building opening a single
book locally. This is nice because you can make changes to the css and
js files and they'll be automatically synced into place. And parcel will
cause them to be reloaded on the fly. This works *perfectly* with css
and seems to cause some flickering with js, but otherwise seems really
useful. I did not do this when building all books locally because I
wanted a way to test "production" settings locally and most of the quick
iteration work I've done has been on a single book.

Finally, this has the lovely side effect of speeding up building a
single book when you don't pass `--open` by entirely skipping the
parcel step. This is a thing some folks do locally to check on the
validity of changes locally before opening a PR. And the tests do it a
whole lot.