Move toc style to its own file
This moves and *slightly* simplifies the style for the table of
contents. It is *super* nesty, mostly because that is how the old styles
were laid out. We expect to rework the toc at some point so it isn'
worth doing any major surgury to it now. But isolating it seems good.