Drop guide.css
Removes our use of `guide.css` which is a css file maintained by the web
team for use with the docs. This merges all of the CSS that it contained
into our css. It tweaks a few small things that I saw while working on
it, like:

1. The screen-size dependant styles for the nav header didn't also apply
to the footer. Now they do.
2. We had a rule that attempted to override the font of the headers but
it wasn't taking. This comments it out so there isn't any confusion and
calls it out for future investigation.

It also factors out some more css into files, specifically the styles
for headers, the "rtpcontainer", and the nav. The "rtpcontainer" is a
little mysterious to me so at the moment I've preserved all of the
styles that it shipped with even though they aren't all used. I bet they
are used in a less than obvious way involving afer-the-fact changes that
don't happen when you render locally. We'll have to get a better handle
of those eventually, but for now we let them stay a mystery.

Closes #1067