Let folks chose their favorite language (#1085)
Allows folks to pick the language they'd like to see instead of "Console"
from a drop down under every "Console" snippet. The drop down includes
every language for which we have alternatives and doesn't show up in
books that don't have alaternative languages configured for "Console".

This is really step one of exposing this feature. There are a few more
things to do:

1. Add a text "bubble" above the warning icon explaining why it shows up.
It usually shows up because the user has picked a language that doesn't
have any examples for that particular snippet.
2. Prevent the "jumping" around that happens when we switch examples.
This is caused by the examples not having the same length.
3. Clean up the "funny" case when a user has a preference for a language
that doesn't show up in the book.
4. Probably performance improvements.
5. Almost certainly style polishing.